+++ tests/integration/test_signed_tt.py
@@ -105,7 +105,7 @@ class TestSignedWithOrWithoutTt(TestBaseSwHeadersAndAttributes):
         assert len(spans) == 2
         span_server = spans[1]
         span_client = spans[0]
-        assert span_server.name == "/test_trace/"
+        assert span_server.name == "GET /test_trace/"
         assert span_server.kind == trace_api.SpanKind.SERVER
         assert span_client.name == "GET"
         assert span_client.kind == trace_api.SpanKind.CLIENT
@@ -265,7 +265,7 @@ class TestSignedWithOrWithoutTt(TestBaseSwHeadersAndAttributes):
         assert len(spans) == 2
         span_server = spans[1]
         span_client = spans[0]
-        assert span_server.name == "/test_trace/"
+        assert span_server.name == "GET /test_trace/"
         assert span_server.kind == trace_api.SpanKind.SERVER
         assert span_client.name == "GET"
         assert span_client.kind == trace_api.SpanKind.CLIENT