+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/solarwinds/apm-python/compare/rel-1.7.0...HEAD)
 
-## [1.7.0.0](https://github.com/solarwinds/apm-python/releases/tag/rel-1.7.0) - 2024-05-15
+## [1.7.0](https://github.com/solarwinds/apm-python/releases/tag/rel-1.7.0) - 2024-05-16
 
 ### Changed
 - Fixed log duplication ([#357](https://github.com/solarwinds/apm-python/pull/357))
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed debug line ([#360](https://github.com/solarwinds/apm-python/pull/360))
 - Sampler inits with APM-proto exporter ([#366](https://github.com/solarwinds/apm-python/pull/366))
 - Fixed APM-proto export for non-scalar attribute types ([#367](https://github.com/solarwinds/apm-python/pull/367))
+
+### Changed
+- OpenTelemetry API/SDK and instrumentation 1.25.0/0.46b0 ([#364](https://github.com/solarwinds/apm-python/pull/364))
 
 ## [1.6.0](https://github.com/solarwinds/apm-python/releases/tag/rel-1.6.0) - 2024-04-02